+++ src/components/events/CommitCommentEvent.vue
@@ -1,6 +1,17 @@
 <template>
   <div class="github-event" style="padding-bottom: 0.5em !important;">
-    <octicon class="event-octicon" name="comment" />
+    <svg
+      class="event-octicon"
+      xmlns="http://www.w3.org/2000/svg"
+      width="16"
+      height="16"
+      viewBox="0 0 16 16"
+    >
+      <path
+        fill-rule="evenodd"
+        d="M14 1H2c-.55 0-1 .45-1 1v8c0 .55.45 1 1 1h2v3.5L7.5 11H14c.55 0 1-.45 1-1V2c0-.55-.45-1-1-1zm0 9H7l-2 2v-2H2V2h12v8z"
+      />
+    </svg>
     <div style="display: inline-block;" class="event-text" v-if="event">
       <a style="text-transform: capitalize;">{{ event.actor.login }}</a>
       <b>