+++ src/components/events/DeleteEvent.vue
@@ -1,6 +1,17 @@
 <template>
   <div class="github-event">
-    <octicon class="event-octicon" name="trashcan" />
+    <svg
+      class="event-octicon"
+      xmlns="http://www.w3.org/2000/svg"
+      width="16"
+      height="16"
+      viewBox="0 0 12 16"
+    >
+      <path
+        fill-rule="evenodd"
+        d="M11 2H9c0-.55-.45-1-1-1H5c-.55 0-1 .45-1 1H2c-.55 0-1 .45-1 1v1c0 .55.45 1 1 1v9c0 .55.45 1 1 1h7c.55 0 1-.45 1-1V5c.55 0 1-.45 1-1V3c0-.55-.45-1-1-1zm-1 12H3V5h1v8h1V5h1v8h1V5h1v8h1V5h1v9zm1-10H2V3h9v1z"
+      />
+    </svg>
     <div style="display: inline-block;" class="event-text" v-if="event">
       <a style="text-transform: capitalize;">{{ event.actor.login }}</a>
       <b>