+++ src/components/events/WatchEvent.vue
@@ -1,6 +1,17 @@
 <template>
   <div class="github-event">
-    <octicon class="event-octicon octicon octicon-star dashboard-event-icon" name="star" />
+    <svg
+      class="event-octicon"
+      xmlns="http://www.w3.org/2000/svg"
+      width="16"
+      height="16"
+      viewBox="0 0 14 16"
+    >
+      <path
+        fill-rule="evenodd"
+        d="M14 6l-4.9-.64L7 1 4.9 5.36 0 6l3.6 3.26L2.67 14 7 11.67 11.33 14l-.93-4.74L14 6z"
+      />
+    </svg>
     <div style="display: inline-block;" class="event-text" v-if="event">
       <a style="text-transform: capitalize;">{{ event.actor.login }}</a>
       <b>
@@ -21,7 +32,7 @@ import { fromNow } from '../../utils/FormatTools'
 
 export default {
   name: 'watch-event',
-   props: {
+  props: {
     event: { required: true }
   },
   methods: {